% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{data}
\name{actualizeEnv}
\alias{actualizeEnv}
\title{Declare a new environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
actualizeEnv
}
\description{
when package is loaded, create new environment to store needed variables
}
\keyword{datasets}
