% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{assign_csv}
\alias{assign_csv}
\title{Read an CSV file and assign to environment,}
\usage{
assign_csv(file_name, envir = actualizeEnv)
}
\arguments{
\item{file_name}{the location of data in your local project}

\item{envir}{where to look for the object}
}
\description{
Read an CSV file and assign to environment,
}
