% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{assign_excel}
\alias{assign_excel}
\title{Read an Excel file and assign to environment,}
\usage{
assign_excel(file_name, envir = actualizeEnv)
}
\arguments{
\item{file_name}{the location of data in your local project}

\item{envir}{where to look for the object}
}
\description{
Read an Excel file and assign to environment,
}
