% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{assign_google_meta}
\alias{assign_google_meta}
\title{Assign Google Drive metadata to a Named value}
\usage{
assign_google_meta(key, envir = actualizeEnv)
}
\arguments{
\item{key}{Something that identifies the file of interest on your
Google Drive. Can be a name or path, a file id or URL marked with
as_id(), or a dribble}

\item{envir}{where to look for the object}
}
\description{
Assign Google Drive metadata to a Named value
}
