% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{assign_tibbles}
\alias{assign_tibbles}
\title{Create tibbles of all files in data-raw}
\usage{
assign_tibbles(file_dir = "~/actualize/data-raw")
}
\arguments{
\item{file_dir}{The place where downloaded files are located.}
}
\description{
Create tibbles of all files in data-raw
}
