% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{download_most_recent_bucket_file}
\alias{download_most_recent_bucket_file}
\title{Parse PAWS bucket contents for most recent file}
\usage{
download_most_recent_bucket_file(
  local_dest_path = "~/actualize/data-raw/",
  file_name = "local1",
  env_file = ".env-data-science"
)
}
\arguments{
\item{local_dest_path}{The place where the file should be downloaded.}

\item{file_name}{The name of the flat you are looking for in the S3 bucket.}

\item{env_file}{The location of your data science environment variable file.}
}
\description{
This function grabs the most recent file in the S3 bucket named in the
`I360_S3_BUCKET` environment variable. It stores the file in the team's
Google Drive and also keeps a local copy for immediate analysis.
}
