% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{extract_match}
\alias{extract_match}
\title{Find variables from a `.env` file}
\usage{
extract_match(line, match)
}
\arguments{
\item{line}{lines of textt read from a `.env` file}

\item{match}{An integer vector giving the position in table
of the first match if there is a match, otherwise nomatch.}
}
\description{
an internal function from the dotenv package
}
