% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{get_max_file_location}
\alias{get_max_file_location}
\title{Parse PAWS bucket contents for most recent file}
\usage{
get_max_file_location(bucket_contents)
}
\arguments{
\item{bucket_contents}{a list returned from the `list_buckets` method in
PAWS S3 tool.}
}
\description{
Parse PAWS bucket contents for most recent file
}
