% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ignore_comments}
\alias{ignore_comments}
\title{Ignore comments in an environment file}
\usage{
ignore_comments(lines)
}
\arguments{
\item{lines}{lines of textt read from a `.env` file}
}
\description{
an internal function from the dotenv package
}
