% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{ignore_empty_lines}
\alias{ignore_empty_lines}
\title{Ignore empty lines in an environment file}
\usage{
ignore_empty_lines(lines)
}
\arguments{
\item{lines}{lines of textt read from a `.env` file}
}
\description{
an internal function from the dotenv package
}
