% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{import_actualize_data}
\alias{import_actualize_data}
\title{Import all actualize data from various sources}
\usage{
import_actualize_data(
  env_file = ".env-data-science",
  file_dir = "~/actualize/data-raw"
)
}
\arguments{
\item{env_file}{The location of your data science environment variable file.}

\item{file_dir}{the designated location of local actualize data. In the production
repository, this is the `data-raw/` directory. This directory should be added to
your .gitignore file as well.}
}
\description{
Import all actualize data from various sources
}
