% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_data}
\alias{load_data}
\title{Download several named objects from Google Drive}
\usage{
load_data(keys, ...)
}
\arguments{
\item{keys}{a vector of object names (given as a character string).}

\item{...}{other arguments to be passed on to `googledrive::drive_auth()`.
In particular, the token parameter for that function should be passed to any
use of this function in production since the function is currently designed
to equire "out of band" authentication.}
}
\description{
Download several named objects from Google Drive
}
