% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_env_vars}
\alias{load_env_vars}
\title{Load environment variables from the specified file}
\usage{
load_env_vars(file = ".env-data-science")
}
\arguments{
\item{file}{The location of your data science environment variable file.}
}
\description{
Load variables defined in the given file, as environment variables, and
also return a vector of variable names or keys.
}
