% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_google_drive_file}
\alias{load_google_drive_file}
\title{Download a named object from Google Drive}
\usage{
load_google_drive_file(obj = "dbg1", envir = actualizeEnv)
}
\arguments{
\item{obj}{An object name (given as a character string).}

\item{envir}{where to look for the object}
}
\description{
Download a named object from Google Drive
}
