% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{load_metadata}
\alias{load_metadata}
\title{Download file metadata from Google Drive}
\usage{
load_metadata(keys, ...)
}
\arguments{
\item{keys}{a vector of something that identifies the file of
interest on your Google Drive. Can be a name or path, a file
id or URL marked with as_id(), or a dribble.}

\item{...}{other arguments to be passed on to `googledrive::drive_auth()`.
In particular, the token parameter for that function should be passed to any
use of this function in production since the function is currently designed
to equire "out of band" authentication.}
}
\description{
Download file metadata from Google Drive
}
