% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{parse_dot_line}
\alias{parse_dot_line}
\title{Find variables from a `.env` file}
\usage{
parse_dot_line(line, envir)
}
\arguments{
\item{line}{lines of textt read from a `.env` file}

\item{envir}{where to look for the object}
}
\description{
an internal function from the dotenv package
}
