% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{set_project_user_token}
\alias{set_project_user_token}
\title{Set token cache from googledrive::drive_auth()}
\usage{
set_project_user_token(secret_cache = ".secrets")
}
\arguments{
\item{secret_cache}{The file path for a secrets file stored to disk.}
}
\description{
Set token cache from googledrive::drive_auth()
}
